refactor: 모임 수요 댓글 도메인 패키지 분리
