+++ main/src/main/java/org/sopt/makers/crew/main/entity/meetingdemandcomment/MeetingDemandCommentRepository.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.crew.main.entity.meetingdemand;
+package org.sopt.makers.crew.main.entity.meetingdemandcomment;
 
 import java.util.List;
 import java.util.Optional;
