+++ main/src/main/java/org/sopt/makers/crew/main/meetingdemandcomment/v2/dto/response/MeetingDemandCommentWriterDto.java
@@ -1,4 +1,4 @@
-package org.sopt.makers.crew.main.meetingdemand.v2.dto.response;
+package org.sopt.makers.crew.main.meetingdemandcomment.v2.dto.response;
 
 import io.swagger.v3.oas.annotations.media.Schema;
 import lombok.AllArgsConstructor;
