+++ main/src/main/java/org/sopt/makers/crew/main/meetingdemandcomment/v2/service/MeetingDemandCommentV2ServiceImpl.java
@@ -1,13 +1,13 @@
-package org.sopt.makers.crew.main.meetingdemand.v2.service;
+package org.sopt.makers.crew.main.meetingdemandcomment.v2.service;
 
 import java.util.List;
 
 import org.sopt.makers.crew.main.entity.meetingdemand.MeetingDemand;
-import org.sopt.makers.crew.main.entity.meetingdemand.MeetingDemandComment;
-import org.sopt.makers.crew.main.entity.meetingdemand.MeetingDemandCommentLike;
-import org.sopt.makers.crew.main.entity.meetingdemand.MeetingDemandCommentLikeRepository;
-import org.sopt.makers.crew.main.entity.meetingdemand.MeetingDemandCommentProfile;
-import org.sopt.makers.crew.main.entity.meetingdemand.MeetingDemandCommentRepository;
+import org.sopt.makers.crew.main.entity.meetingdemandcomment.MeetingDemandComment;
+import org.sopt.makers.crew.main.entity.meetingdemandcomment.MeetingDemandCommentLike;
+import org.sopt.makers.crew.main.entity.meetingdemandcomment.MeetingDemandCommentLikeRepository;
+import org.sopt.makers.crew.main.entity.meetingdemandcomment.MeetingDemandCommentProfile;
+import org.sopt.makers.crew.main.entity.meetingdemandcomment.MeetingDemandCommentRepository;
 import org.sopt.makers.crew.main.entity.meetingdemand.MeetingDemandRepository;
 import org.sopt.makers.crew.main.entity.user.User;
 import org.sopt.makers.crew.main.entity.user.UserRepository;
@@ -15,14 +15,15 @@ import org.sopt.makers.crew.main.global.pagination.dto.PageMetaDto;
 import org.sopt.makers.crew.main.global.pagination.dto.PageOptionsDto;
 import org.sopt.makers.crew.main.global.util.MentionSecretStringRemover;
 import org.sopt.makers.crew.main.global.util.Time;
-import org.sopt.makers.crew.main.meetingdemand.v2.dto.query.MeetingDemandCommentV2GetCommentsQueryDto;
-import org.sopt.makers.crew.main.meetingdemand.v2.dto.request.MeetingDemandCommentV2CreateCommentBodyDto;
-import org.sopt.makers.crew.main.meetingdemand.v2.dto.request.MeetingDemandCommentV2MentionUserInCommentRequestDto;
-import org.sopt.makers.crew.main.meetingdemand.v2.dto.response.MeetingDemandCommentDto;
-import org.sopt.makers.crew.main.meetingdemand.v2.dto.response.MeetingDemandCommentV2CreateCommentResponseDto;
-import org.sopt.makers.crew.main.meetingdemand.v2.dto.response.MeetingDemandCommentV2GetCommentsResponseDto;
-import org.sopt.makers.crew.main.meetingdemand.v2.dto.response.MeetingDemandCommentV2SwitchCommentLikeResponseDto;
-import org.sopt.makers.crew.main.meetingdemand.v2.dto.response.MeetingDemandCommentV2UpdateCommentResponseDto;
+import org.sopt.makers.crew.main.meetingdemand.v2.service.MeetingDemandPageNormalizer;
+import org.sopt.makers.crew.main.meetingdemandcomment.v2.dto.query.MeetingDemandCommentV2GetCommentsQueryDto;
+import org.sopt.makers.crew.main.meetingdemandcomment.v2.dto.request.MeetingDemandCommentV2CreateCommentBodyDto;
+import org.sopt.makers.crew.main.meetingdemandcomment.v2.dto.request.MeetingDemandCommentV2MentionUserInCommentRequestDto;
+import org.sopt.makers.crew.main.meetingdemandcomment.v2.dto.response.MeetingDemandCommentDto;
+import org.sopt.makers.crew.main.meetingdemandcomment.v2.dto.response.MeetingDemandCommentV2CreateCommentResponseDto;
+import org.sopt.makers.crew.main.meetingdemandcomment.v2.dto.response.MeetingDemandCommentV2GetCommentsResponseDto;
+import org.sopt.makers.crew.main.meetingdemandcomment.v2.dto.response.MeetingDemandCommentV2SwitchCommentLikeResponseDto;
+import org.sopt.makers.crew.main.meetingdemandcomment.v2.dto.response.MeetingDemandCommentV2UpdateCommentResponseDto;
 import org.springframework.data.domain.Page;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
